протестировал на запуск в c#

+++ CPP/include/common/FrameGpu.h
@@ -10,7 +10,7 @@ class FrameGpu
 public:
     FrameGpu() = default;
 
-    FrameGpu(T* image, const int width,const int height,const uint64_t timestamp, const int channel)
+    FrameGpu(T* image, const int width,const int height,const uint32_t timestamp, const int channel)
     {
         if(!image || width <= 0 || height <= 0 || timestamp <= 0 || channel <= 0)
             throw std::runtime_error("[FrameGpu::Ctr char, int, int, uint64_t, int] Fail input parameters ptr:" +
@@ -55,14 +55,14 @@ public:
     T* ImagePtr()  const { return _images; }
 
     unsigned int GetFulSize() const { return _width * _height * _channels; }
-    void SetTimestamp(uint64_t timestamp)
+    void SetTimestamp(uint32_t timestamp)
     {
         _timestamp = timestamp;
     };
 
 private:
     T * _images = nullptr;
-    uint64_t _timestamp = 0;
+    uint32_t _timestamp = 0;
     int _width = -1;
     int _height = -1;
     int _channels = -1;
